+++ sdk/resiliencemanagement/azure-resourcemanager-resiliencemanagement/src/main/java/com/azure/resourcemanager/resiliencemanagement/models/GoalResourceProperties.java
@@ -10,7 +10,6 @@ import com.azure.json.JsonSerializable;
 import com.azure.json.JsonToken;
 import com.azure.json.JsonWriter;
 import java.io.IOException;
-import java.util.List;
 
 /**
  * Definition of goal assignment property.
@@ -23,44 +22,16 @@ public final class GoalResourceProperties implements JsonSerializable<GoalResour
     private String resourceArmId;
 
     /*
-     * Flag which depicts whether the Arm resource is excluded for high availability recommendation.
+     * Zonal resiliency posture (participation, attestation, exclusion reason, and user confirmations) for the Arm
+     * resource.
      */
-    private ExclusionState highAvailabilityGoalParticipation;
+    private ResiliencyProperties zonalResiliency;
 
     /*
-     * Flag which depicts whether the Arm resource is manually attested for high availability recommendation.
+     * Regional resiliency posture (participation, attestation, exclusion reason, and user confirmations) for the Azure
+     * resource.
      */
-    private AttestationState highAvailabilityAttestationStatus;
-
-    /*
-     * Flag which depicts whether the Arm resource is excluded for disaster recovery recommendation.
-     */
-    private ExclusionState disasterRecoveryGoalParticipation;
-
-    /*
-     * Flag which depicts whether the Arm resource is manually attested for disaster recovery recommendation.
-     */
-    private AttestationState disasterRecoveryAttestationStatus;
-
-    /*
-     * Reason for exclusion from high availability goals.
-     */
-    private ExclusionReason exclusionReasonForHighAvailabilityGoals;
-
-    /*
-     * Reason for exclusion from disaster recovery goals.
-     */
-    private ExclusionReason exclusionReasonForDisasterRecoveryGoals;
-
-    /*
-     * List of user confirmations for high availability solutions.
-     */
-    private List<UserConfirmationForHighAvailabilityItem> userConfirmationForHighAvailability;
-
-    /*
-     * List of service groups of which this resource is memberof.
-     */
-    private List<ServiceGroupMembership> serviceGroupMemberships;
+    private ResiliencyProperties regionalResiliency;
 
     /*
      * Provisioning state
@@ -94,143 +65,47 @@ public final class GoalResourceProperties implements JsonSerializable<GoalResour
     }
 
     /**
-     * Get the highAvailabilityGoalParticipation property: Flag which depicts whether the Arm resource is excluded for
-     * high availability recommendation.
+     * Get the zonalResiliency property: Zonal resiliency posture (participation, attestation, exclusion reason, and
+     * user confirmations) for the Arm resource.
      * 
-     * @return the highAvailabilityGoalParticipation value.
+     * @return the zonalResiliency value.
      */
-    public ExclusionState highAvailabilityGoalParticipation() {
-        return this.highAvailabilityGoalParticipation;
+    public ResiliencyProperties zonalResiliency() {
+        return this.zonalResiliency;
     }
 
     /**
-     * Set the highAvailabilityGoalParticipation property: Flag which depicts whether the Arm resource is excluded for
-     * high availability recommendation.
+     * Set the zonalResiliency property: Zonal resiliency posture (participation, attestation, exclusion reason, and
+     * user confirmations) for the Arm resource.
      * 
-     * @param highAvailabilityGoalParticipation the highAvailabilityGoalParticipation value to set.
+     * @param zonalResiliency the zonalResiliency value to set.
      * @return the GoalResourceProperties object itself.
      */
-    public GoalResourceProperties
-        withHighAvailabilityGoalParticipation(ExclusionState highAvailabilityGoalParticipation) {
-        this.highAvailabilityGoalParticipation = highAvailabilityGoalParticipation;
+    public GoalResourceProperties withZonalResiliency(ResiliencyProperties zonalResiliency) {
+        this.zonalResiliency = zonalResiliency;
         return this;
     }
 
     /**
-     * Get the highAvailabilityAttestationStatus property: Flag which depicts whether the Arm resource is manually
-     * attested for high availability recommendation.
+     * Get the regionalResiliency property: Regional resiliency posture (participation, attestation, exclusion reason,
+     * and user confirmations) for the Azure resource.
      * 
-     * @return the highAvailabilityAttestationStatus value.
+     * @return the regionalResiliency value.
      */
-    public AttestationState highAvailabilityAttestationStatus() {
-        return this.highAvailabilityAttestationStatus;
+    public ResiliencyProperties regionalResiliency() {
+        return this.regionalResiliency;
     }
 
     /**
-     * Set the highAvailabilityAttestationStatus property: Flag which depicts whether the Arm resource is manually
-     * attested for high availability recommendation.
+     * Set the regionalResiliency property: Regional resiliency posture (participation, attestation, exclusion reason,
+     * and user confirmations) for the Azure resource.
      * 
-     * @param highAvailabilityAttestationStatus the highAvailabilityAttestationStatus value to set.
+     * @param regionalResiliency the regionalResiliency value to set.
      * @return the GoalResourceProperties object itself.
      */
-    public GoalResourceProperties
-        withHighAvailabilityAttestationStatus(AttestationState highAvailabilityAttestationStatus) {
-        this.highAvailabilityAttestationStatus = highAvailabilityAttestationStatus;
+    public GoalResourceProperties withRegionalResiliency(ResiliencyProperties regionalResiliency) {
+        this.regionalResiliency = regionalResiliency;
         return this;
-    }
-
-    /**
-     * Get the disasterRecoveryGoalParticipation property: Flag which depicts whether the Arm resource is excluded for
-     * disaster recovery recommendation.
-     * 
-     * @return the disasterRecoveryGoalParticipation value.
-     */
-    public ExclusionState disasterRecoveryGoalParticipation() {
-        return this.disasterRecoveryGoalParticipation;
-    }
-
-    /**
-     * Set the disasterRecoveryGoalParticipation property: Flag which depicts whether the Arm resource is excluded for
-     * disaster recovery recommendation.
-     * 
-     * @param disasterRecoveryGoalParticipation the disasterRecoveryGoalParticipation value to set.
-     * @return the GoalResourceProperties object itself.
-     */
-    public GoalResourceProperties
-        withDisasterRecoveryGoalParticipation(ExclusionState disasterRecoveryGoalParticipation) {
-        this.disasterRecoveryGoalParticipation = disasterRecoveryGoalParticipation;
-        return this;
-    }
-
-    /**
-     * Get the disasterRecoveryAttestationStatus property: Flag which depicts whether the Arm resource is manually
-     * attested for disaster recovery recommendation.
-     * 
-     * @return the disasterRecoveryAttestationStatus value.
-     */
-    public AttestationState disasterRecoveryAttestationStatus() {
-        return this.disasterRecoveryAttestationStatus;
-    }
-
-    /**
-     * Set the disasterRecoveryAttestationStatus property: Flag which depicts whether the Arm resource is manually
-     * attested for disaster recovery recommendation.
-     * 
-     * @param disasterRecoveryAttestationStatus the disasterRecoveryAttestationStatus value to set.
-     * @return the GoalResourceProperties object itself.
-     */
-    public GoalResourceProperties
-        withDisasterRecoveryAttestationStatus(AttestationState disasterRecoveryAttestationStatus) {
-        this.disasterRecoveryAttestationStatus = disasterRecoveryAttestationStatus;
-        return this;
-    }
-
-    /**
-     * Get the exclusionReasonForHighAvailabilityGoals property: Reason for exclusion from high availability goals.
-     * 
-     * @return the exclusionReasonForHighAvailabilityGoals value.
-     */
-    public ExclusionReason exclusionReasonForHighAvailabilityGoals() {
-        return this.exclusionReasonForHighAvailabilityGoals;
-    }
-
-    /**
-     * Get the exclusionReasonForDisasterRecoveryGoals property: Reason for exclusion from disaster recovery goals.
-     * 
-     * @return the exclusionReasonForDisasterRecoveryGoals value.
-     */
-    public ExclusionReason exclusionReasonForDisasterRecoveryGoals() {
-        return this.exclusionReasonForDisasterRecoveryGoals;
-    }
-
-    /**
-     * Get the userConfirmationForHighAvailability property: List of user confirmations for high availability solutions.
-     * 
-     * @return the userConfirmationForHighAvailability value.
-     */
-    public List<UserConfirmationForHighAvailabilityItem> userConfirmationForHighAvailability() {
-        return this.userConfirmationForHighAvailability;
-    }
-
-    /**
-     * Set the userConfirmationForHighAvailability property: List of user confirmations for high availability solutions.
-     * 
-     * @param userConfirmationForHighAvailability the userConfirmationForHighAvailability value to set.
-     * @return the GoalResourceProperties object itself.
-     */
-    public GoalResourceProperties withUserConfirmationForHighAvailability(
-        List<UserConfirmationForHighAvailabilityItem> userConfirmationForHighAvailability) {
-        this.userConfirmationForHighAvailability = userConfirmationForHighAvailability;
-        return this;
-    }
-
-    /**
-     * Get the serviceGroupMemberships property: List of service groups of which this resource is memberof.
-     * 
-     * @return the serviceGroupMemberships value.
-     */
-    public List<ServiceGroupMembership> serviceGroupMemberships() {
-        return this.serviceGroupMemberships;
     }
 
     /**
@@ -249,16 +124,8 @@ public final class GoalResourceProperties implements JsonSerializable<GoalResour
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
         jsonWriter.writeStringField("resourceArmId", this.resourceArmId);
-        jsonWriter.writeStringField("highAvailabilityGoalParticipation",
-            this.highAvailabilityGoalParticipation == null ? null : this.highAvailabilityGoalParticipation.toString());
-        jsonWriter.writeStringField("highAvailabilityAttestationStatus",
-            this.highAvailabilityAttestationStatus == null ? null : this.highAvailabilityAttestationStatus.toString());
-        jsonWriter.writeStringField("disasterRecoveryGoalParticipation",
-            this.disasterRecoveryGoalParticipation == null ? null : this.disasterRecoveryGoalParticipation.toString());
-        jsonWriter.writeStringField("disasterRecoveryAttestationStatus",
-            this.disasterRecoveryAttestationStatus == null ? null : this.disasterRecoveryAttestationStatus.toString());
-        jsonWriter.writeArrayField("userConfirmationForHighAvailability", this.userConfirmationForHighAvailability,
-            (writer, element) -> writer.writeJson(element));
+        jsonWriter.writeJsonField("zonalResiliency", this.zonalResiliency);
+        jsonWriter.writeJsonField("regionalResiliency", this.regionalResiliency);
         return jsonWriter.writeEndObject();
     }
 
@@ -280,33 +147,10 @@ public final class GoalResourceProperties implements JsonSerializable<GoalResour
 
                 if ("resourceArmId".equals(fieldName)) {
                     deserializedGoalResourceProperties.resourceArmId = reader.getString();
-                } else if ("highAvailabilityGoalParticipation".equals(fieldName)) {
-                    deserializedGoalResourceProperties.highAvailabilityGoalParticipation
-                        = ExclusionState.fromString(reader.getString());
-                } else if ("highAvailabilityAttestationStatus".equals(fieldName)) {
-                    deserializedGoalResourceProperties.highAvailabilityAttestationStatus
-                        = AttestationState.fromString(reader.getString());
-                } else if ("disasterRecoveryGoalParticipation".equals(fieldName)) {
-                    deserializedGoalResourceProperties.disasterRecoveryGoalParticipation
-                        = ExclusionState.fromString(reader.getString());
-                } else if ("disasterRecoveryAttestationStatus".equals(fieldName)) {
-                    deserializedGoalResourceProperties.disasterRecoveryAttestationStatus
-                        = AttestationState.fromString(reader.getString());
-                } else if ("exclusionReasonForHighAvailabilityGoals".equals(fieldName)) {
-                    deserializedGoalResourceProperties.exclusionReasonForHighAvailabilityGoals
-                        = ExclusionReason.fromString(reader.getString());
-                } else if ("exclusionReasonForDisasterRecoveryGoals".equals(fieldName)) {
-                    deserializedGoalResourceProperties.exclusionReasonForDisasterRecoveryGoals
-                        = ExclusionReason.fromString(reader.getString());
-                } else if ("userConfirmationForHighAvailability".equals(fieldName)) {
-                    List<UserConfirmationForHighAvailabilityItem> userConfirmationForHighAvailability
-                        = reader.readArray(reader1 -> UserConfirmationForHighAvailabilityItem.fromJson(reader1));
-                    deserializedGoalResourceProperties.userConfirmationForHighAvailability
-                        = userConfirmationForHighAvailability;
-                } else if ("serviceGroupMemberships".equals(fieldName)) {
-                    List<ServiceGroupMembership> serviceGroupMemberships
-                        = reader.readArray(reader1 -> ServiceGroupMembership.fromJson(reader1));
-                    deserializedGoalResourceProperties.serviceGroupMemberships = serviceGroupMemberships;
+                } else if ("zonalResiliency".equals(fieldName)) {
+                    deserializedGoalResourceProperties.zonalResiliency = ResiliencyProperties.fromJson(reader);
+                } else if ("regionalResiliency".equals(fieldName)) {
+                    deserializedGoalResourceProperties.regionalResiliency = ResiliencyProperties.fromJson(reader);
                 } else if ("provisioningState".equals(fieldName)) {
                     deserializedGoalResourceProperties.provisioningState
                         = ProvisioningState.fromString(reader.getString());