+++ sdk/resiliencemanagement/azure-resourcemanager-resiliencemanagement/src/main/java/com/azure/resourcemanager/resiliencemanagement/models/RegionalDrillProperties.java
@@ -51,6 +51,16 @@ public final class RegionalDrillProperties extends DrillProperties {
      * {@inheritDoc}
      */
     @Override
+    public RegionalDrillProperties
+        withGoalAssignmentProperties(GoalAssignmentPropertiesOfDrill goalAssignmentProperties) {
+        super.withGoalAssignmentProperties(goalAssignmentProperties);
+        return this;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
     public RegionalDrillProperties withDrillAssetProperties(AssetPropertiesOfDrill drillAssetProperties) {
         super.withDrillAssetProperties(drillAssetProperties);
         return this;
@@ -87,13 +97,35 @@ public final class RegionalDrillProperties extends DrillProperties {
      * {@inheritDoc}
      */
     @Override
+    public RegionalDrillProperties
+        withHealthModelMonitoringProperties(HealthModelMonitoringProperties healthModelMonitoringProperties) {
+        super.withHealthModelMonitoringProperties(healthModelMonitoringProperties);
+        return this;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public RegionalDrillProperties withSliMonitoringProperties(SliMonitoringProperties sliMonitoringProperties) {
+        super.withSliMonitoringProperties(sliMonitoringProperties);
+        return this;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
         jsonWriter.writeJsonField("recoveryPlanProperties", recoveryPlanProperties());
+        jsonWriter.writeJsonField("goalAssignmentProperties", goalAssignmentProperties());
         jsonWriter.writeJsonField("drillAssetProperties", drillAssetProperties());
         jsonWriter.writeJsonField("chaosResourceProperties", chaosResourceProperties());
         jsonWriter.writeStringField("rbacSetupMode", rbacSetupMode() == null ? null : rbacSetupMode().toString());
         jsonWriter.writeJsonField("monitoringProperties", monitoringProperties());
+        jsonWriter.writeJsonField("healthModelMonitoringProperties", healthModelMonitoringProperties());
+        jsonWriter.writeJsonField("sliMonitoringProperties", sliMonitoringProperties());
         jsonWriter.writeStringField("drillType", this.drillType == null ? null : this.drillType.toString());
         return jsonWriter.writeEndObject();
     }
@@ -121,6 +153,9 @@ public final class RegionalDrillProperties extends DrillProperties {
                 } else if ("recoveryPlanProperties".equals(fieldName)) {
                     deserializedRegionalDrillProperties
                         .withRecoveryPlanProperties(RecoveryPlanPropertiesOfDrill.fromJson(reader));
+                } else if ("goalAssignmentProperties".equals(fieldName)) {
+                    deserializedRegionalDrillProperties
+                        .withGoalAssignmentProperties(GoalAssignmentPropertiesOfDrill.fromJson(reader));
                 } else if ("drillAssetProperties".equals(fieldName)) {
                     deserializedRegionalDrillProperties
                         .withDrillAssetProperties(AssetPropertiesOfDrill.fromJson(reader));
@@ -150,6 +185,12 @@ public final class RegionalDrillProperties extends DrillProperties {
                 } else if ("monitoringProperties".equals(fieldName)) {
                     deserializedRegionalDrillProperties
                         .withMonitoringProperties(MonitoringPropertiesOfDrill.fromJson(reader));
+                } else if ("healthModelMonitoringProperties".equals(fieldName)) {
+                    deserializedRegionalDrillProperties
+                        .withHealthModelMonitoringProperties(HealthModelMonitoringProperties.fromJson(reader));
+                } else if ("sliMonitoringProperties".equals(fieldName)) {
+                    deserializedRegionalDrillProperties
+                        .withSliMonitoringProperties(SliMonitoringProperties.fromJson(reader));
                 } else if ("errorDetails".equals(fieldName)) {
                     deserializedRegionalDrillProperties.withErrorDetails(ManagementError.fromJson(reader));
                 } else if ("drillType".equals(fieldName)) {